adding file headers and  updating other syntax
