+++ firestore/FirestoreSwiftUIExample/Extensions/Firestore+Extension.swift
@@ -2,6 +2,20 @@
 //  Firestore+Extension.swift
 //  FirestoreSwiftUIExample
 //
+//  Copyright (c) 2021 Google Inc.
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//  http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+//
 
 import Firebase
 
@@ -43,15 +57,12 @@ extension Firestore {
       }
 
       let batch = self.batch()
-      //TODO: guard let user = Auth.auth().currentUser else { continue }
       var average: Float = 0
       for _ in 0 ..< 10 {
         let rating = Int(arc4random_uniform(5) + 1)
         average += Float(rating) / 10
         let text = rating > 3 ? "good" : "food was too spicy"
 
-        //TODO: userID: user.uid,
-        //TODO: username: user.displayName ?? "Anonymous",
         let review = Review(
           rating: rating,
           userID: "1234567890",
