+++ firestore/FirestoreSwiftUIExample/Views/PriceView.swift
@@ -2,6 +2,20 @@
 //  PriceView.swift
 //  FirestoreSwiftUIExample
 //
+//  Copyright (c) 2021 Google Inc.
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//  http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+//
 
 import SwiftUI
 
@@ -16,18 +30,10 @@ struct PriceView: View {
   }
 
   func priceString(from price: Int) -> String {
-    let priceText: String
-    switch price {
-    case 1:
-      priceText = "$"
-    case 2:
-      priceText = "$$"
-    case 3:
-      priceText = "$$$"
-    case _:
+    if ![1,2,3].contains(price) {
       fatalError("price must be between one and three")
     }
 
-    return priceText
+    return String(repeating: "$", count: price)
   }
 }
