+++ firestore/FirestoreSwiftUIExample/Views/RestaurantListView.swift
@@ -1,12 +1,26 @@
 //
-//  ContentView.swift
+//  RestaurantListView.swift
 //  FirestoreSwiftUIExample
+//
+//  Copyright (c) 2021 Google Inc.
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//  http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
 //
 
 import SwiftUI
 import Firebase
 
-struct ContentView: View {
+struct RestaurantListView: View {
   @ObservedObject var restaurantListViewModel = RestaurantListViewModel()
 
   var body: some View {
@@ -16,7 +30,7 @@ struct ContentView: View {
           RestaurantItemView(restaurant: restaurant)
         }
       }
-      .navigationBarTitle("Friendly Eats", displayMode: .inline)
+      .navigationBarTitle("Friendly Eats")
       .onAppear() {
         restaurantListViewModel.subscribe()
       }
@@ -40,14 +54,15 @@ struct ContentView: View {
   }
 }
 
+// Work In Progress: Stub for detailed view of restaurants
 struct RestaurantDetailView: View {
   var body: some View {
     Text("Detail View Placeholder")
   }
 }
 
-struct ContentView_Previews: PreviewProvider {
+struct RestaurantListView_Previews: PreviewProvider {
   static var previews: some View {
-    ContentView(restaurantListViewModel: RestaurantListViewModel())
+    RestaurantListView(restaurantListViewModel: RestaurantListViewModel())
   }
 }
